# packages/openapi/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Poveroh API",
        "version": "1.0.0",
        "description": "OpenAPI definition for the Poveroh v1 API"
    },
    "servers": [
        {
            "url": "http://localhost:3001/v1",
            "description": "API Poveroh Localhost v1"
        },
        {
            "url": "http://api.poveroh.local/v1",
            "description": "API Poveroh Localhost proxy v1"
        },
        {
            "url": "http://api.poveroh.com/v1",
            "description": "API Poveroh Production v1"
        }
    ],
    "components": {
        "schemas": {
...
            "StringFilter": {
                "type": "object",
                "properties": {
                    "equals": {
                        "type": "string"
                    },
                    "contains": {
                        "type": "string"
                    }
                }
            },
...
            "FilterOptions": {
                "type": "object",
                "properties": {
                    "skip": {
                        "type": "integer",
                        "minimum": 0,
                        "default": 0
                    },
                    "take": {
                        "type": "integer",
                        "minimum": 0,
                        "exclusiveMinimum": true,
                        "default": 20
                    },
                    "sortBy": {
                        "type": "string",
                        "default": "createdAt"
                    },
                    "sortOrder": {
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "default": "desc"
                    }
                }
            },
...
            "FinancialAccountTypeEnum": {
                "type": "string",
                "enum": [
                    "ONLINE_BANK",
                    "BANK_ACCOUNT",
                    "CIRCUIT",
                    "DEPOSIT_BANK",
                    "BROKER",
                    "WALLET",
                    "CASH",
                    "CREDIT_CARD",
                    "OTHER"
                ]
            },
...
            "AppearanceModeEnum": {
                "type": "string",
                "enum": [
                    "LOGO",
                    "ICON"
                ]
            },
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            false
                        ],
                        "description": "Always false for error responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Error message describing what went wrong"
                    },
                    "error": {
                        "nullable": true,
                        "description": "Optional additional error details or context"
                    }
                },
                "required": [
                    "success",
                    "message"
                ]
            },
...
            "GetSubscriptionListResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SubscriptionDataResponse"
                        },
                        "description": "Response data"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "GetSubscriptionResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/SubscriptionDataResponse"
                            },
                            {
                                "description": "Response data"
                            }
                        ]
                    }
                },
                "required": [
                    "success"
                ]
            },
            "CreateSubscriptionRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "currency": {
                        "$ref": "#/components/schemas/CurrencyEnum"
                    },
                    "appearanceMode": {
                        "$ref": "#/components/schemas/AppearanceModeEnum"
                    },
                    "appearanceLogoIcon": {
                        "type": "string"
                    },
                    "firstPayment": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "cycleNumber": {
                        "type": "number"
                    },
                    "cyclePeriod": {
                        "type": "string"
                    },
                    "rememberPeriod": {
                        "$ref": "#/components/schemas/RememberPeriodEnum"
                    },
                    "financialAccountId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "isEnabled": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "title",
                    "description",
                    "amount",
                    "currency",
                    "appearanceMode",
                    "appearanceLogoIcon",
                    "firstPayment",
                    "cycleNumber",
                    "cyclePeriod",
                    "rememberPeriod",
                    "financialAccountId",
                    "isEnabled"
                ]
            },
            "CreateSubscriptionMultipartRequest": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/CreateSubscriptionRequest"
                    },
                    "file": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary",
                            "description": "Optional file upload"
                        }
                    }
                },
                "required": [
                    "file"
                ]
            },
            "CreateSubscriptionResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/SubscriptionDataResponse"
                            },
                            {
                                "description": "Response data"
                            }
                        ]
                    }
                },
                "required": [
                    "success"
                ]
            },
            "UpdateSubscriptionRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "currency": {
                        "$ref": "#/components/schemas/CurrencyEnum"
                    },
                    "appearanceMode": {
                        "$ref": "#/components/schemas/AppearanceModeEnum"
                    },
                    "appearanceLogoIcon": {
                        "type": "string"
                    },
                    "firstPayment": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "cycleNumber": {
                        "type": "number"
                    },
                    "cyclePeriod": {
                        "type": "string"
                    },
                    "rememberPeriod": {
                        "$ref": "#/components/schemas/RememberPeriodEnum"
                    },
                    "financialAccountId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "isEnabled": {
                        "type": "boolean"
                    }
                }
            },
            "UpdateSubscriptionResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "nullable": true,
                        "description": "Response data"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "DeleteSubscriptionResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "nullable": true,
                        "description": "Response data"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "SubscriptionParamsId": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "id"
                ]
            },
            "SubscriptionFilters": {
                "type": "object",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/SubscriptionParamsId"
                    },
                    "title": {
                        "$ref": "#/components/schemas/StringFilter"
                    },
                    "description": {
                        "$ref": "#/components/schemas/StringFilter"
                    },
                    "type": {
                        "$ref": "#/components/schemas/FinancialAccountTypeEnum"
                    }
                },
                "additionalProperties": {
                    "anyOf": [
                        {
                            "type": "string"
                        },
                        {
                            "$ref": "#/components/schemas/StringFilter"
                        },
                        {
                            "$ref": "#/components/schemas/FinancialAccountTypeEnum"
                        }
                    ]
                }
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/subscriptions": {
            "get": {
                "tags": [
                    "Subscription"
                ],
                "operationId": "getSubscriptions",
                "summary": "Get all subscriptions",
                "description": "Retrieve a list of all subscriptions associated with the user",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/components/schemas/SubscriptionFilters"
                        },
                        "required": false,
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "$ref": "#/components/schemas/FilterOptions"
                        },
                        "required": false,
                        "name": "options",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of subscriptions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetSubscriptionListResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subscription not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Subscription"
                ],
                "operationId": "createSubscription",
                "summary": "Create subscription",
                "description": "Create a new subscription with the provided data",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "description": "Subscription data to create",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSubscriptionMultipartRequest"
                            },
                            "encoding": {
                                "data": {
                                    "contentType": "application/json"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateSubscriptionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subscription"
                ],
                "operationId": "deleteSubscriptions",
                "summary": "Delete all subscriptions",
                "description": "Delete all subscriptions associated with the user",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscriptions deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteSubscriptionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/subscriptions/{id}": {
            "get": {
                "tags": [
                    "Subscription"
                ],
                "operationId": "getSubscriptionById",
                "summary": "Get subscription by ID",
                "description": "Retrieve a specific subscription by its ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetSubscriptionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subscription not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Subscription"
                ],
                "operationId": "updateSubscription",
                "summary": "Update subscription",
                "description": "Update an existing subscription with the provided data",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "description": "Subscription data to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSubscriptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateSubscriptionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Financial account not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subscription"
                ],
                "operationId": "deleteSubscription",
                "summary": "Delete subscription",
                "description": "Delete an existing subscription with the provided ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteSubscriptionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subscription not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}